% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHelpers.R
\name{calcTransRate}
\alias{calcTransRate}
\title{Calculate total transmission rate}
\usage{
calcTransRate(beta, S, I, N)
}
\arguments{
\item{beta}{The transmission coefficient (can be thought of as the per capita
contact rate times the probability of trasnmission given contact between a
susceptible and an infectious individual) (numeric, positive).}

\item{S}{The number of susceptible individuals in the population at a given
point in time (numeric, positive).}

\item{I}{The number of infectious individuals in the population at a given
point in time (numeric, positive).}

\item{N}{The total population size to be used in the calculation of the
transmission rate (numeric, positive). Can be either the population size at
the beginning of the epidemic (as in \code{simpleSim}) or the population
size at the time point of interest (i.e., the initial population size minus
the number of individuals who have died).}
}
\description{
\code{calcTransRate} calculates the total transmission rate at a given point
in time
}
