% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHelpers.R
\name{GenerateIDs}
\alias{GenerateIDs}
\title{Generate IDs for all population members}
\usage{
GenerateIDs()
}
\value{
Returns a random 5-letter string.
}
\description{
\code{generateIDs} creates random IDs for each individual in the simulated
population. This function is used internally by \code{simpleSim}.
}
