% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHelpers.R
\name{get_inf}
\alias{get_inf}
\title{Determine infectious period}
\usage{
get_inf(number, shape, rate)
}
\arguments{
\item{number}{The number of random number draws to be made (integer).}

\item{shape}{The shape parameter of the Gamma distribution to draw from
(numeric).}

\item{rate}{The rate parameter of the Gamma distribution to draw from
(numeric).}
}
\description{
\code{get_inf} selects \code{number} random values for the infectious period
from a Gamma distribution with the specified shape and rate parameters.
}
\details{
Currently, \code{get_inf} is a wrapper for the \code{rgamma} function from
the \code{stats} package. The intent is to implement this more generically in
the future to allow use of distributions other than Gamma.
}
