% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSim.R
\name{simpleSim}
\alias{simpleSim}
\title{Simple SIR simulation}
\usage{
simpleSim(
  inc_mean_param,
  inf_mean_param,
  R_0,
  case_fatality,
  pop_size,
  inc_shape_param = 1,
  inf_shape_param = 1,
  initial_cases = 1,
  max_time = 365,
  seed = 20200129
)
}
\arguments{
\item{inc_mean_param}{Mean of the incubation period distribution}

\item{inf_mean_param}{Mean of the infectious period}

\item{R_0}{Basic reproductive number}

\item{case_fatality}{Proportion of cases that die}

\item{pop_size}{Size of the population}

\item{inc_shape_param}{Shape parameter of the incubation period distribution
(default = 1)}

\item{inf_shape_param}{Shape parameter of the infectious period distribution
(default = 1)}

\item{initial_cases}{Initial number of cases (default = 1)}

\item{max_time}{Maximum time for the simulation to run (default = 365)}

\item{seed}{Seed for the random number generator (default = 20200129)}
}
\value{
Returns a list with two dataframes: \code{pop}, the full population
  on which the simulation was run, and \code{cases} a preliminary line list
  (perfect reporting)

  TODO: Add more detail
}
\description{
\code{simpleSim} runs a stochastic SIR model with gamma-distributed waiting
times.
}
\details{
TODO: Add more details later
}
